Thing work... I had a longer message at one point...
Basically I had to reset because wonky things are happening in my
frontend directory, and considering I don't think any of it should be
happening because it's not my domain, I reset...

As for the backend, it's all working now. The only issue it has is DUE
TO THE FRONTEND, and their needing to access other things in the web
directory, which means we can't just serve the html directory.

If you start up the server and go to /html, you'll see everything
working. Provided you're using dartium. Otherwise you might not see
everything working...