Get a list of all files in the music dir!
With the one issue of the files currently need to be something that uses
ID3 tags. So far as I know, only MP3 does that...

If you navigate to /songs, you should get a json array of json objects
of the songs and some of their ID3 info. The big ones are
artist/album/name.