Added a metadata display.
Whenever Don finishes being able send us our song info, we should be able to
display it. This should work if all the json data is correct, but I'm not sure
if it is. I'm not sure if `/current` is the best path for it. We can change
it if there is a better one.