controller.go can detect MPD changes!
So, MPD apparently has a nice observer-esque thing going for it, and we
get access to that using our gompd package. Huzzah!

Basically, at present, if you run controller.go, and run any command
using mpc, you'll be shown that *something* has changed. I can, and
probably should, look into doing things with that knowledge. But we have
it, and it's exciting!