And look at our little "server.go"!
Because it's tiny now. Which is AWESOME! I'm just really happy that it's
tiny. And when I say tiny, I mean TINY. 16 lines, including four empty
lines, and three lines with just closing punctuation.

It's nice.

Also, it means that pretty much everything is typed now. I personally
think that's a good thing, because everything now only knows what it
needs to, and nothing more. As in, the HTTP server doesn't know anything
about how we talk to MPD.

As for how we talk to MPD, I'm not sure I described that yet ...

Where we previously would attempt sending the same command a couple
times, and keep the connection open forever (and ever), we now create a
new connection each time we want to talk to MPD. It's suprisingly quick.
I was not expecting it to be anywhere near as fast as it is. At least on
my machine, there isn't any preceivable difference from before.

On to the files, which at some point I'll describe more accurately and
fully in a piece of our deliverables:

Everything is now inside of a separate turbo-wookie directory. Inside we
have four files: TBHandler.go; TBMPDClient.go; turbo-wookie.go; and
yaml.go.

TBHandler contains all of the HTTP stuff. And really just the HTTP
stuff. A TBHandler is a type of HttpHandler, and we cover up the fact
that next to no implementation went into building it by just using a
mux.Handler as a data member and writing a couple "pass-through"
functions.

TBMPDClient takes care of all our MPD interactions. It has a similar
interface to the gompd/mpd package, but currently only has the functions
we use (plus a couple others). It could use with a quick clean up...

turbo-wookie has a couple generic error types for us. Thuroughly
uninteresting.

yaml takes care of all our YAML parsing. Because the package we're using
doesn't just store it as a map of strings to strings in memory (which is
all we really need), this goes through, grabs only the key/values we
need, and ignores everything else. It's only one function, no types, and
just returns a map[string]string of the keys we want.