Removed private methods.
I was going to add in private variables next, but it turns out that private
things in dart only apply to things outside of the library. So it would have
no benifit to us at this point. I feel like it would just be annoying to press
the `_` key every time we want to access a private thing, so it's not worth
it.