We're closer to being able to use this in a real browser.
It still doesn't work, but now it at least shows our polymer stuff.
To build it, run `pub build` (you may need to run it on the pubspec.yaml file).
With the dart editor, you can just right click on `index.html` and select
`Build Polymer App`. It places all of the compiled down code in the `build`
folder.