Though, hey, go has a built in formatter, let's use that...
And I did. All of our code has been go formated, using `go fmt`. Because
it comes with the language.

Used a custom command, because the default one likes slightly different
formatting.

If you, yes you, are considering doing this, run the following on both
the backend directory, and the backend/turbo-wookie directory:

    gofmt -tabs=false -tabwidth=2 -s -w *.go

What that does is tell gofmt that we want to use two space hard tabs
(instead of hard tabs), to simplify the code if it can (optimizations
are nice), and overwrite the current code (if you don't include the -w
flag it'll just output the new code to stdout, which is kinda
useles...).