Made MPD query the library for info now...
Mostly I've been making some changes to the MPD client library because
the original author only implemented the things he needed/wanted (which
is fine). Considering we need other things, I've added them to a fork,
and I'm waiting for approval on my pull request.

The main reason for this was I had a song that started causing errors in
our MP3 parser (the problem was with the file), which made me think
there *had* to be a way to make MPD give us that information without
using some external library (especially considering that it's all in the
MPD database).

Anyways, everything should now return in the same format, the MPD
format, instead of having a couple things sending over multiple formats.