Added a hook to get to automatically build Turbo Wookie on pull.
To use it, place the file in the folder `git-hooks` in your `.git/hooks`
folder with the name of `post-merge`, and no extension.

You also must have python installed.

You may or may not have to be sure you have permissions to run the script
before it works for you.

Finally, You may or may not have to change the top line of the file to be
your python path. I'm not exactly sure how that part works.