When the library gets updated in MPD, update frontend.
Previously, the frontend was looking for a change saying the "library"
was updated, but because we just send whatever MPD tells us has changed,
and there is no "library" subsystem, the frontend never updates.

Instead, we need to watch for "database". It now does.

But, to be fair, the only time MPD will update is when it launches, or
when someone manually tells it to update, outside of Turbo Wookie.