Added Artist, Albums and Songs tabs to the library.
So far, the Songs tab is the only one that works correctly (for the most part).
The Artist tab shows a list of artists, but you can't do anything with it yet
(and it looks pretty bad). The Album tab doesn't do anything at all so far.

The add to playlist button is broken on the Songs tab, but that should be an
easy fix (should).