Fixed an issue with the title div click listener.
It slowed down so much because I was adding a new click listener every time
we loaded an artist page. So when we clicked it, it loaded an artist page,
and told the title div to add another click listener(not set like I was
thinking). So sometimes it was even loading the wrong page as well because
it was finished doing an old click listener last (or something like that).

Anyway it's fast now and it works correctly.