Made some small changes to the title div.
It now shows the album you are looking at as well. In the future, we should
probably have a subtitle div just because it's strange to be able to click
on the album title and it brings you back to the artist. It would probably
look better if they were styled differently as well.