Our Albums tab works!
So we now send the client the albums in a slightly different way.
Before, we just sent a list of Strings, now we send a Map of Lists of
Strings. The Map's key is the artist name, and the List's values are the
album names. That way the frontend knows what album belongs to what
artist.