Updated code for name change and made links work on the playlist.
I forgot to update the actual code when I change the name of the player.

I also made the links for the playlist work. You can click on any artist
and go to their album page. And any album, and go to the album's page.
This includes the current song and any song in the playlist.