Searching works. For the most part.
Searching for artists works well.

A few issues:
If we add songs, it'll probably crash the page because there are so many
of them. This should probably be delt with somehow.

Albums need to know their artists from the server, before we can really
display this, but a placeholder is there for now.

The paper element that I wanted to use is broken (yay), so I commented
out the code for it and left in a boring input element for now.